fixes for fourRouters example

+++ src/generic_env.h
@@ -613,7 +613,7 @@ struct oppsimt_in_addr
         union {
                 struct { u_char s_b1,s_b2,s_b3,s_b4; } S_un_b;
                 struct { u_short s_w1,s_w2; } S_un_w;
-                u_long S_addr;
+                u32 S_addr;
         } S_un;
 #define s_addr  S_un.S_addr
 #define s_host  S_un.S_un_b.s_b2
